Adds message command to application cli besides commander
