+++ examples/pos-mainchain/src/commands/message/decrypt.ts
@@ -1,0 +1,1 @@
+export { MessageDecryptCommand } from 'lisk-commander';
