+++ examples/pos-mainchain/src/commands/message/encrypt.ts
@@ -1,0 +1,1 @@
+export { MessageEncryptCommand } from 'lisk-commander';
