+++ examples/pos-mainchain/src/commands/message/sign.ts
@@ -1,0 +1,1 @@
+export { MessageSignCommand } from 'lisk-commander';
