+++ examples/pos-mainchain/src/commands/message/verify.ts
@@ -1,0 +1,1 @@
+export { MessageVerifyCommand } from 'lisk-commander';
